@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { FormGroup, FormBuilder, Validators, AbstractControl} from '@angular/forms';
+import { FormGroup, FormBuilder, FormControl, Validators, AbstractControl} from '@angular/forms';
 
-import 'rxjs/add/operator/do';
+import { tap } from 'rxjs/operators';
 
 import { RadioOption } from '../shared/radio/radio-option.model';
 import { OrderService } from './order.service';
@@ -36,8 +36,10 @@ export class OrderComponent implements OnInit {
               private formBuilder: FormBuilder) { }
 
   ngOnInit() {
-    this.orderForm = this.formBuilder.group({
-      name: this.formBuilder.control('', [Validators.required, Validators.minLength(5)]),
+    this.orderForm = new FormGroup({
+      name: new FormControl('', {
+        validators: [Validators.required, Validators.minLength(5)]
+      }),
       email: this.formBuilder.control('', [Validators.required, Validators.pattern(this.emailPattern)]),
       emailConfirmation: this.formBuilder.control('', [Validators.required, Validators.pattern(this.emailPattern)]),
       address: this.formBuilder.control('', [Validators.required, Validators.minLength(5)]),
@@ -45,7 +47,7 @@ export class OrderComponent implements OnInit {
       optionalAddress: this.formBuilder.control(''),
       paymentOption: this.formBuilder.control('', [Validators.required])
     }, {
-      validator: OrderComponent.equalsTo
+      validators: [OrderComponent.equalsTo], updateOn: 'blur'
     })
   }
 
@@ -92,10 +94,11 @@ export class OrderComponent implements OnInit {
       .map((item:CartItem) => new OrderItem(item.quantity, item.menuItem.id))
     
     this.orderService.checkOrder(order)
-      .do((orderId: string) => {
-        this.orderId = orderId
-      })
-      .subscribe((orderId: string) => {
+      .pipe(
+        tap((orderId: string) => {
+          this.orderId = orderId
+        })
+      ).subscribe((orderId: string) => {
         this.router.navigate(['/order-summary'])
         this.orderService.clear()
       })
