@@ -4,7 +4,7 @@ import { ActivatedRoute } from '@angular/router';
 
 import { RestaurantsService } from '../../restaurants/restaurants.service';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'mt-reviews',
