@@ -2,11 +2,11 @@ import { Component, OnInit } from '@angular/core';
 import { trigger, state, style, transition, animate } from '@angular/animations';
 import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 
-
 import { Restaurant } from './restaurant/restaurant.model';
 import { RestaurantsService } from './restaurants.service';
 
-import 'rxjs/add/operator/switchMap';
+import { Observable, from } from 'rxjs';
+import { tap, debounceTime, distinctUntilChanged, switchMap, catchError } from 'rxjs/operators';
 
 @Component({
   selector: 'mt-restaurants',
@@ -45,9 +45,15 @@ export class RestaurantsComponent implements OnInit {
     })
 
     this.searchControl.valueChanges
-      .switchMap(searchTerm => 
-        this.restaurantService.restaurants(searchTerm))
-      .subscribe(restaurants => this.restaurants = restaurants)
+      .pipe(
+        debounceTime(500),
+        distinctUntilChanged(),
+        switchMap(searchTerm => 
+          this.restaurantService
+            .restaurants(searchTerm)
+            .pipe(catchError(error => from([])))
+        )
+      ).subscribe(restaurants => this.restaurants = restaurants)
 
     this.restaurantService.restaurants()
       .subscribe(restaurants => this.restaurants = restaurants);
