@@ -2,9 +2,8 @@ import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
 import { Router, NavigationEnd } from "@angular/router";
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/filter';
+import { Observable } from 'rxjs';
+import { tap, filter } from "rxjs/operators";
 
 import { MEAT_API } from '../../app.api';
 import { User } from "./user.model";
@@ -17,7 +16,7 @@ export class LoginService {
 
   constructor(private http: HttpClient,
               private router: Router) {
-    this.router.events.filter(e => e instanceof NavigationEnd)
+    this.router.events.pipe(filter(e => e instanceof NavigationEnd))
                       .subscribe( (e: NavigationEnd) => this.lastUrl = e.url)
   }
 
@@ -27,7 +26,7 @@ export class LoginService {
 
   login(email: string, password: string): Observable<User> {
     return this.http.post<User>(`${MEAT_API}/login`, {email: email, password: password})
-                    .do(user => this.user = user)
+                    .pipe(tap(user => this.user = user))
   }
 
   handleLogin(path: string = this.lastUrl) {
